Migrated logger to helper classes instead of a small separate file
